refactor: refactor for vuex-sync-store
