+++ README.md
@@ -1,4 +1,4 @@
-# sync-store
+# vue-sync-store
 
 [![CircleCI](https://circleci.com/gh/galenwarren/sync-store/tree/master.svg?style=svg)](https://circleci.com/gh/galenwarren/sync-store/tree/master)
 [![tested with jest](https://img.shields.io/badge/tested_with-jest-99424f.svg)](https://github.com/facebook/jest)
@@ -6,7 +6,7 @@
 [![semantic-release](https://img.shields.io/badge/%20%20%F0%9F%93%A6%F0%9F%9A%80-semantic--release-e10079.svg)](https://github.com/semantic-release/semantic-release)
 [![Commitizen friendly](https://img.shields.io/badge/commitizen-friendly-brightgreen.svg)](http://commitizen.github.io/cz-cli/)
 
-Pluggable store syncing utilities
+Pluggable store syncing utilities for Vue/Vuex
 
 [Documentation](https://galenwarren.github.io/sync-store/)
 
@@ -16,6 +16,7 @@ Use `npm run commit` to commit using [Commitizen](http://commitizen.github.io/cz
 
 Notes:
 
-* If used with Vue, then the [vue-rx](https://github.com/vuejs/vue-rx) plugin must be installed
-* If used with Vuex, must install mutations (more detail needed here)
+* Requires that the [vue-rx](https://github.com/vuejs/vue-rx) plugin be installed
+* Must install mutations for Vuex (more detail needed here)
 * For Nuxt, show how to use with asyncData
+* Roadmap for plugin development
