+++ src/index.js
@@ -1,3 +1,4 @@
 export * from "./observable";
 export * from "./transforms";
+export * from "./mixin";
 export * from "./utils";
